feat: RQES SDK ui, implementing the QTSP selection bottom sheet
Now set QRSPData class for the list within EudiRQESUiConfig to hold, in order to contain the QTSP name additionally, set internal modifier where needed, providing a list of QTSPData in the TestRQESApplication for the RQES SDK to use.
Bottom bar is now placed in the SignDocumentScreen by applying it to the ContentScreen bottomBar parameter, using the navigationBarsPadding operator to move bar above device button navigation row.

Signed-off-by: Christos Kaitatzis <christos.kaitatzis@netcompany.com>